Import parquet only inside function
The parquet Python library is not available on Debian, so we needed to
run tests using `pytest tests` (without the `rows` directory). This is
bad, since there are doctests and just having pytest importing everthing
inside `rows` would be a good "smoke test".
So the import to parquet was moved to inside `import_from_parquet`
function. Since the tests for parquet are skipped if the library is not
installed, the command `pytest` will not having problem collecting on
all directories.